+++ token-client/src/main/java/com/sap/cloud/security/ias/client/IasTenantHostCacheConfiguration.java
@@ -3,7 +3,7 @@
  * <p>
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.cloud.security.ias.tokenflows;
+package com.sap.cloud.security.ias.client;
 
 import jakarta.annotation.Nonnull;
 import java.time.Duration;
