+++ token-client/src/test/java/com/sap/cloud/security/ias/tokenflows/IasTokenFlowsTest.java
@@ -5,6 +5,10 @@
  */
 package com.sap.cloud.security.ias.tokenflows;
 
+import com.sap.cloud.security.ias.client.IasDefaultEndpoints;
+import com.sap.cloud.security.ias.client.IasTenantHostCacheConfiguration;
+import com.sap.cloud.security.ias.client.IasTenantHostResolver;
+
 import com.sap.cloud.security.config.ClientCredentials;
 import com.sap.cloud.security.config.ClientIdentity;
 import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
